+++ .env
@@ -1,6 +1,6 @@
-AUTH0_DOMAIN=https://booking-api/
-AUTH0_CLIENT_ID=uhG6sTlvPPqVI20BiVb4NjA84PBTUPug
-AUTH0_CLIENT_SECRET=5tQhg6jJrqiku25WGa0WZd9YDbQNZKsY80W1-0ZiDC4U74e_xKOJ6ann6ilMbSCC
+AUTH0_DOMAIN=https://booking-web-app-api/
+AUTH0_CLIENT_ID=z19frJEwGXkFQdzavbdnZohsW8Yjjr3c
+AUTH0_CLIENT_SECRET=8V0Ws5mrqinKCJFOcApSnHSECkFYqY19JIOjNavJ8Ep68a4kAicDVuRiA-H34xti
 AUTH0_CONNECTION=your-connection
-AUTH0_CALLBACK_URL=https://your-app/callback
+AUTH0_CALLBACK_URL=http://localhost:5002/signin-auth0
 SESSION_SECRET=your-session-secret
